The assertNotificationWhenLetter method validates a templated letter and a precompiled letter. Only address_line_1 is set for a precompiled letter. This update will allow the check to work for both types of letters.
When we test the getAllNotifications method we get a list of Notifications back, and we validate the first one in the list. So we can check any type of notification, it could be that we either have ignored this error before or that a precompiled letter was never the first in the list. Either way this will resolve the issue.

Fix assertNotificationWhenLetter

Remove .DS_Store file